ps-docparser/requirements.txt
# PDF 처리
pdfplumber>=0.11.0,<1.0.0
pdf2image>=1.17.0,<2.0.0
Pillow>=11.0.0,<12.0.0

# AI / OCR 엔진
google-generativeai>=0.8.0,<1.0.0

# 설정 & 파싱
python-dotenv>=1.0.0,<2.0.0
beautifulsoup4>=4.12.0,<5.0.0
lxml>=5.0.0,<6.0.0

# Excel 출력
openpyxl>=3.1.0,<4.0.0

# (선택) OCR 엔진 — 설치되어 있지 않으면 해당 엔진만 비활성화
# pytesseract>=0.3.10,<1.0.0
# mistralai>=0.4.0,<1.0.0
# zai>=0.1.0,<1.0.0
